+++ libs/environment/src/utils/validate-environment.ts
@@ -1,18 +1,5 @@
-import type { ZodIssue } from 'zod';
 import z from 'zod';
-import type { Environment } from '../types';
-import { compileErrors } from './compile-errors';
-
-export enum Networks {
-  CUSTOM = 'CUSTOM',
-  SANDBOX = 'SANDBOX',
-  TESTNET = 'TESTNET',
-  STAGNET1 = 'STAGNET1',
-  STAGNET3 = 'STAGNET3',
-  DEVNET = 'DEVNET',
-  MAINNET = 'MAINNET',
-  MIRROR = 'MIRROR',
-}
+import { Networks } from '../types';
 
 const schemaObject = {
   VEGA_URL: z.optional(z.string()),
@@ -34,7 +21,7 @@ const schemaObject = {
           [env]: z.optional(z.string()),
         }),
         {}
-      )
+      ) as Record<Networks, z.ZodOptional<z.ZodString>>
     )
     .strict({
       message: `All keys in NX_VEGA_NETWORKS must represent a valid environment: ${Object.keys(
@@ -54,23 +41,6 @@ const schemaObject = {
   ETH_WALLET_MNEMONIC: z.optional(z.string()),
 };
 
-export const ENV_KEYS = Object.keys(schemaObject) as Array<
-  keyof typeof schemaObject
->;
-
-const compileIssue = (issue: ZodIssue) => {
-  switch (issue.code) {
-    case 'invalid_type':
-      return `NX_${issue.path[0]} is invalid, received "${issue.received}" instead of: ${issue.expected}`;
-    case 'invalid_enum_value':
-      return `NX_${issue.path[0]} is invalid, received "${
-        issue.received
-      }" instead of: ${issue.options.join(' | ')}`;
-    default:
-      return issue.message;
-  }
-};
-
 export const envSchema = z.object(schemaObject).refine(
   (data) => {
     return !(!data.VEGA_URL && !data.VEGA_CONFIG_URL);
@@ -80,15 +50,3 @@ export const envSchema = z.object(schemaObject).refine(
       'Must provide either NX_VEGA_CONFIG_URL or NX_VEGA_URL in the environment.',
   }
 );
-
-export const validateEnvironment = (
-  environment: Environment
-): string | undefined => {
-  try {
-    envSchema.parse(environment);
-    return undefined;
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  } catch (err: any) {
-    return compileErrors('Error processing the vega app environment', err);
-  }
-};
